requirements.txt
# Core dependencies for TikTok Ads MCP Server
requests>=2.32.3
python-dotenv>=1.0.0

# MCP (Model Context Protocol) server
mcp>=1.9.0

# Data handling (lightweight)
pandas>=2.2.0